make all task specs in bpmn processes bpmn tasks
